Adding the Java, (Gradle), instance
This instance is a stand alone that can work without modifications, (apart of the need for the private key of course), and it can work with both java 10 and 11.
Few warnings are thrown on build because parts of the code are deprecated. There are solutions for that but we decided to keep our solution as close to the original examples as possible.